vertex_ai_agent/requirements.txt
google-adk==1.14.0
langchain-community==0.3.27
google-cloud-aiplatform==1.50.0
requests==2.31.0
python-dotenv==1.0.0